Removed the hashing of the large files which fixed the issue with uploading them to the server. Added WaitForFile which will wait for about 5 minutes for the file to be transferred to the local folder before trying to stream it to the server.

+++ MOVEit-TransferApp/Hubs/UploadNotificationHub.cs
@@ -23,7 +23,7 @@ namespace MOVEit_TransferApp.Hubs
             await base.OnDisconnectedAsync(exception);
         }
 
-        public async Task FolderNotification(string fileName, int size)
+        public async Task FolderNotification(string fileName, long size)
         {
             if (_connetionId != null)
             {
